requirements.txt
Flask==3.0.2
flasgger==0.9.7.1
Flask-PyMongo==2.3.0
pymongo==4.6.2
Werkzeug==3.0.1
flask-cors==4.0.0
markupsafe==2.1.3
python-dotenv
